CohortDiagnostics-package {CohortDiagnostics}	R Documentation

CohortDiagnostics: Diagnostics for OHDSI Cohorts

Description

CohortDiagnostics is an R utility package for the development and evaluation of phenotype algorithms for OMOP CDM compliant data sets. This package provides a standard, end to end, set of analytics for understanding patient capture including data generation and result exploration through an R Shiny interface. Analytics computed include cohort characteristics, record counts, index event misclassification, captured observation windows and basic incidence proportions for age, gender and calendar year. Through the identification of errors, CohortDiagnostics enables the comparison of multiple candidate cohort definitions across one or more data sources, facilitating reproducible research.

Author(s)

Maintainer: Gowtham Rao rao@ohdsi.org

Authors:

Martijn Schuemie schuemie@ohdsi.org

Patrick Ryan

James Weaver

Jamie Gilbert

Other contributors:

Observational Health Data Science and Informatics [copyright holder]

See Also

Useful links:

https://ohdsi.github.io/CohortDiagnostics

https://github.com/OHDSI/CohortDiagnostics

Report bugs at https://github.com/OHDSI/CohortDiagnostics/issues

[Package CohortDiagnostics version 3.1.2 Index]