createDiagnosticsExplorerZip {CohortDiagnostics}	R Documentation

Create publishable shiny zip

Description

A utility designed for creating a published zip of a shiny app with an sqlite database. Designed for sharing projects on servers like data.ohdsi.org.

Takes the shiny code from the R project and adds an sqlite file to a zip archive. Uncompressed cohort diagnostics sqlite databases can become large very quickly.

Usage

createDiagnosticsExplorerZip(
  outputZipfile = file.path(getwd(), "DiagnosticsExplorer.zip"),
  sqliteDbPath = "MergedCohortDiagnosticsData.sqlite",
  shinyDirectory = system.file(file.path("shiny", "DiagnosticsExplorer"), package =
    "CohortDiagnostics"),
  overwrite = FALSE
)

Arguments

outputZipfile
The output path for the zip file

sqliteDbPath
Merged Cohort Diagnostics sqlitedb created with createMergedResultsFile

shinyDirectory
(optional) Path to the location where the shiny code is stored. By default, this is the package root

overwrite
If the zip file already exists, overwrite it?

[Package CohortDiagnostics version 3.1.2 Index]