createMergedResultsFile {CohortDiagnostics}	R Documentation

Merge Shiny diagnostics files into sqlite database

Description

This function combines diagnostics results from one or more databases into a single file. The result is an sqlite database that can be used as input for the Diagnostics Explorer Shiny app.

It also checks whether the results conform to the results data model specifications.

Usage

createMergedResultsFile(
  dataFolder,
  sqliteDbPath = "MergedCohortDiagnosticsData.sqlite",
  overwrite = FALSE,
  tablePrefix = ""
)

Arguments

dataFolder
folder where the exported zip files for the diagnostics are stored. Use the executeDiagnostics function to generate these zip files. Zip files containing results from multiple databases may be placed in the same folder.

sqliteDbPath
Output path where sqlite database is placed

overwrite
(Optional) overwrite existing sqlite lite db if it exists.

tablePrefix
(Optional) string to insert before table names (e.g. "cd_") for database table names

[Package CohortDiagnostics version 3.1.2 Index]