Create the results data model tables on a database server.

Create the results data model tables on a database server.

createResultsDataModel(connection = NULL, connectionDetails = NULL, schema)

Arguments

connection

An object of type connection as created using the connect function in the DatabaseConnector package. Can be left NULL if connectionDetails is provided, in which case a new connection will be opened at the start of the function, and closed when the function finishes.

connectionDetails

An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package. Can be left NULL if connection is provided.

schema

The schema on the postgres server where the tables will be created.

Details

Only PostgreSQL servers are supported.