Rsconnect deploy

Deploy your application to an posit connect platform or shinyapps.io server

Usage

deployPositConnectApp(
  appName,
  appDir = tempfile(),
  sqliteDbPath = "MergedCohortDiagnosticsData.sqlite",
  shinyDirectory = system.file(file.path("shiny", "DiagnosticsExplorer"), package =
    "CohortDiagnostics"),
  connectionDetails = NULL,
  shinyConfigPath = NULL,
  resultsDatabaseSchema = NULL,
  vocabularyDatabaseSchemas = resultsDatabaseSchema,
  tablePrefix = "",
  cohortTableName = "cohort",
  databaseTableName = "database",
  port = 80,
  useRenvironFile = FALSE,
  ...
)

Arguments

appName

string name to call app - should be unique on posit connect server

appDir

optional - directory to use to copy files for deployment. If you use a consistent dir other internal options can change.

sqliteDbPath

Path to merged sqlite file. See createMergedResultsFile to create file.

shinyDirectory

(optional) Directyory shiny app code lives. Use this if you wish to modify the explorer

connectionDetails

An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package, specifying how to connect to the server where the CohortDiagnostics results have been uploaded using the uploadResults function.

shinyConfigPath

Path to shiny yml configuration file (use instead of sqliteDbPath or connectionDetails object)

resultsDatabaseSchema

The schema on the database server where the CohortDiagnostics results have been uploaded.

vocabularyDatabaseSchemas

(optional) A list of one or more schemas on the database server where the vocabulary tables are located. The default value is the value of the resultsDatabaseSchema. We can provide a list of vocabulary schema that might represent different versions of the OMOP vocabulary tables. It allows us to compare the impact of vocabulary changes on Diagnostics. Not supported with an sqlite database.

tablePrefix

(Optional) string to insert before table names (e.g. "cd_") for database table names

cohortTableName

(Optional) if cohort table name differs from the standard - cohort (ignores prefix if set)

databaseTableName

(Optional) if database table name differs from the standard - database (ignores prefix if set)

port

(optional) Only used if runOverNetwork = TRUE.

useRenvironFile

logical - not recommended, store db credentials in .Renviron file

...

other parameters passed to rsconnect::deployApp