getDataMigrator {CohortDiagnostics}	R Documentation

Get database migrations instance

Description

Returns ResultModelManager DataMigrationsManager instance.

Usage

getDataMigrator(connectionDetails, databaseSchema, tablePrefix = "")

Arguments

connectionDetails
DatabaseConnector connection details object

databaseSchema
String schema where database schema lives

tablePrefix
(Optional) Use if a table prefix is used before table names (e.g. "cd_")

Value

Instance of ResultModelManager::DataMigrationManager that has interface for converting existing data models

[Package CohortDiagnostics version 3.1.2 Index]