Compute incidence proportion for a cohort

Returns yearly incidence proportion time series data stratified by age and gender

getIncidenceProportion(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema,
  cohortTable,
  cdmDatabaseSchema,
  firstOccurrenceOnly = TRUE,
  minObservationTime = 365,
  cohortId
)

Arguments

connectionDetails
An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package. Can be left NULL if connection is provided.

connection
An object of type connection as created using the connect function in the DatabaseConnector package. Can be left NULL if connectionDetails is provided, in which case a new connection will be opened at the start of the function, and closed when the function finishes.

cohortDatabaseSchema
Schema name where your cohort table resides. Note that for SQL Server, this should include both the database and schema name, for example 'scratch.dbo'.

cohortTable
Name of the cohort table.

cdmDatabaseSchema
Schema name where your patient-level data in OMOP CDM format resides. Note that for SQL Server, this should include both the database and schema name, for example 'cdm_data.dbo'.

firstOccurrenceOnly
Use only the first occurrence of the cohort per person?

minObservationTime
The minimum amount of observation time required before the occurrence of a cohort entry. This is also used to eliminate immortal time from the denominator.

cohortId
The cohort definition ID used to reference the cohort in the cohort table.

Value

A data frame

Details

Returns a data frame of cohort count, background count, and incidence proportion per 1000 persons of cohort entry with the following stratifications: 1) no stratification, 2) gender stratification, 3) age (10-year) stratification, 4) calendar year and age (10-year) stratification, 5) calendar year and gender stratification, 6) calendar year, age (10-year), and gender stratification with option to save dataframes.