Launch the CohortExplorer Shiny app

Launch the CohortExplorer Shiny app

launchCohortExplorer(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  cohortId,
  sampleSize = 100,
  subjectIds = NULL
)

Arguments

connectionDetails

An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package.

cdmDatabaseSchema

Schema name where your patient-level data in OMOP CDM format resides. Note that for SQL Server, this should include both the database and schema name, for example 'cdm_data.dbo'.

cohortDatabaseSchema

Schema name where your cohort table resides. Note that for SQL Server, this should include both the database and schema name, for example 'scratch.dbo'.

cohortTable

Name of the cohort table.

cohortId

The ID of the cohort.

sampleSize

Number of subjects to sample from the cohort. Ignored if subjectIds is specified.

subjectIds

A vector of subject IDs to view.

Details

Launches a Shiny app that allows the user to explore a cohort of interest.