loadCohortsFromPackage {CohortDiagnostics}	R Documentation

Load Cohort Definitions From A Study Package

Description

Load cohort references for usage in executeDiagnostics.

Usage

loadCohortsFromPackage(
  packageName,
  cohortToCreateFile = "settings/cohortsToCreate.csv",
  cohortIds = NULL,
  errorMessage = NULL
)

Arguments

packageName
The name of the package containing the cohort definitions. Can be left NULL if baseUrl and cohortSetReference have been specified.

cohortToCreateFile
The location of the cohortToCreate file within the package. Is ignored if baseUrl and cohortSetReference have been specified. The cohortToCreateFile must be .csv file that is expected to be read into a dataframe object identical to requirements for cohortSetReference argument. This csv file is expected to be encoded in either ASCII or UTF-8, if not, an error message will be displayed and process stopped.

cohortIds
Optionally, provide a subset of cohort IDs to restrict the diagnostics to.

errorMessage
checkmate assert collection, used internally for error checks

[Package CohortDiagnostics version 2.2.4 Index]