migrateDataModel {CohortDiagnostics}	R Documentation

Migrate Data model

Description

Migrate data from current state to next state

It is strongly advised that you have a backup of all data (either sqlite files, a backup database (in the case you are using a postgres backend) or have kept the csv/zip files from your data generation.

Usage

migrateDataModel(connectionDetails, databaseSchema, tablePrefix = "")

Arguments

connectionDetails
DatabaseConnector connection details object

databaseSchema
String schema where database schema lives

tablePrefix
(Optional) Use if a table prefix is used before table names (e.g. "cd_")

[Package CohortDiagnostics version 3.1.2 Index]