preMergeDiagnosticsFiles {CohortDiagnostics}	R Documentation

Premerge Shiny diagnostics files

Description

This function combines diagnostics results from one or more databases into a single file. The result is a single file that can be used as input for the Diagnostics Explorer Shiny app.

It also checks whether the results conform to the results data model specifications.

Usage

preMergeDiagnosticsFiles(dataFolder, tempFolder = tempdir())

Arguments

dataFolder
folder where the exported zip files for the diagnostics are stored. Use the runCohortDiagnostics function to generate these zip files. Zip files containing results from multiple databases may be placed in the same folder.

tempFolder
A folder on the local file system where the zip files are extracted to. Will be cleaned up when the function is finished. Can be used to specify a temp folder on a drive that has sufficient space if the default system temp space is too limited.

[Package CohortDiagnostics version 2.1.4 Index]