runCohortRelationshipDiagnostics {CohortDiagnostics}	R Documentation

Given a set of cohorts get relationships between the cohorts.

Description

Given a set of cohorts, get temporal relationships between the cohort_start_date of the cohorts.

Usage

runCohortRelationshipDiagnostics(
  connectionDetails = NULL,
  connection = NULL,
  cohortDatabaseSchema = NULL,
  tempEmulationSchema = NULL,
  cohortTable = "cohort",
  targetCohortIds,
  comparatorCohortIds,
  relationshipDays
)

Arguments

connectionDetails
An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package. Can be left NULL if connection is provided.

connection
An object of type connection as created using the connect function in the DatabaseConnector package. Can be left NULL if connectionDetails is provided, in which case a new connection will be opened at the start of the function, and closed when the function finishes.

cohortDatabaseSchema
Schema name where your cohort table resides. Note that for SQL Server, this should include both the database and schema name, for example 'scratch.dbo'.

tempEmulationSchema
Some database platforms like Oracle and Impala do not truly support temp tables. To emulate temp tables, provide a schema with write privileges where temp tables can be created.

cohortTable
Name of the cohort table.

targetCohortIds
A vector of one or more Cohort Ids for use as target cohorts.

comparatorCohortIds
A vector of one or more Cohort Ids for use as feature/comparator cohorts.

relationshipDays
A dataframe with two columns startDay and endDay representing periods of time to compute relationship

[Package CohortDiagnostics version 3.1.2 Index]