Run study diagnostics

Runs the study diagnostics on all (or a subset of) the cohorts instantiated using the ROhdsiWebApi::insertCohortDefinitionSetInPackage function.

runStudyDiagnostics(
  packageName,
  cohortToCreateFile = "settings/CohortsToCreate.csv",
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  tempEmulationSchema = NULL,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  cohortIds = NULL,
  inclusionStatisticsFolder = NULL,
  exportFolder,
  databaseId,
  databaseName,
  databaseDescription,
  runInclusionStatistics = TRUE,
  runIncludedSourceConcepts = TRUE,
  runOrphanConcepts = TRUE,
  runBreakdownIndexEvents = TRUE,
  runIncidenceProportion = TRUE,
  runCohortOverlap = TRUE,
  runCohortCharacterization = TRUE
)

Arguments

packageName
The name of the package containing the cohort definitions

cohortToCreateFile
The location of the cohortToCreate file within the package.

connectionDetails
An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package. Can be left NULL if connection is provided.

connection
An object of type connection as created using the connect function in the DatabaseConnector package. Can be left NULL if connectionDetails is provided, in which case a new connection will be opened at the start of the function, and closed when the function finishes.

cdmDatabaseSchema
Schema name where your patient-level data in OMOP CDM format resides. Note that for SQL Server, this should include both the database and schema name, for example 'cdm_data.dbo'.

tempEmulationSchema
Should be used in Oracle to specify a schema where the user has write privileges for storing temporary tables.

cohortDatabaseSchema
Schema name where your cohort table resides. Note that for SQL Server, this should include both the database and schema name, for example 'scratch.dbo'.

cohortTable
Name of the cohort table.

cohortIds
Optionally, provide a subset of cohort IDs to restrict the diagnostics to.

inclusionStatisticsFolder
The folder where the inclusion rule statistics are stored. Can be left NULL if runInclusionStatistics = FALSE.

exportFolder
The folder where the output will be exported to. If this folder does not exist it will be created.

databaseId
A short string for identifying the database (e.g. 'Synpuf').

databaseName
The full name of the database.

databaseDescription
A short description (several sentences) of the database.

runInclusionStatistics
Generate and export statistic on the cohort incusion rules?

runIncludedSourceConcepts
Generate and export the source concepts included in the cohorts?

runOrphanConcepts
Generate and export potential orphan concepts?

runBreakdownIndexEvents
Generate and export the breakdown of index events?

runIncidenceProportion
Generate and export the cohort incidence proportions?

runCohortOverlap
Generate and export the cohort overlap?

runCohortCharacterization
Generate and export the cohort characterization?