Get a copy of OMOP vocabulary as csv

For a given list of conceptId's get a subset of OMOP vocabulary of these conceptIds. These are written as csv in the export folder

writeOmopvocabularyTables(
  connectionDetails = NULL,
  connection = NULL,
  vocabularyDatabaseSchema = NULL,
  conceptIds = NULL,
<<<<<<< HEAD
  vocabularyTableNames = c("concept", "conceptAncestor", "conceptClass",
    "conceptRelationship", "conceptSynonym", "domain", "relationship", "vocabulary"),
=======
  vocabularyTableNames = c("concept", "domain", "vocabulary", "relationship",
    "conceptClass", "conceptAncestor", "conceptRelationship"),
>>>>>>> master
  exportFolder
)

Arguments

connectionDetails
An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package. Can be left NULL if connection is provided.

connection
An object of type connection as created using the connect function in the DatabaseConnector package. Can be left NULL if connectionDetails is provided, in which case a new connection will be opened at the start of the function, and closed when the function finishes.

vocabularyDatabaseSchema
databaseSchema where the OMOP vocabulary files are located. These are most commonly the same as cdmDatabaseSchema.

conceptIds <<<<<<< HEAD
(optional) A vector of conceptIds to filter the extract of OMOP vocabulary files. If NULL, all conceptIds are extracted.
=======
A vector of conceptIds to extract OMOP vocabulary files.
>>>>>>> master

vocabularyTableNames
(optional) A vector of OMOP vocabulary table names to download.

exportFolder
The folder where the output is exported by Cohort Diagnostics. If this folder does not exist, or does not have the searched file the function will return an error.

Value

NULL. The function writes the vocabulary tables into the export folder as csv.