createResultsDataModel {CohortDiagnostics}	R Documentation

Create the results data model tables on a database server.

Description

Create the results data model tables on a database server.

Usage

createResultsDataModel(
  connectionDetails = NULL,
  databaseSchema,
  tablePrefix = ""
)

Arguments

connectionDetails
DatabaseConnector connectionDetails instance @seealso[DatabaseConnector::createConnectionDetails]

databaseSchema
The schema on the postgres server where the tables will be created.

tablePrefix
(Optional) string to insert before table names (e.g. "cd_") for database table names

Details

Only PostgreSQL servers are supported.

[Package CohortDiagnostics version 3.1.2 Index]