instantiateCohort {CohortDiagnostics}	R Documentation

Instantiate a cohort

Description

This function instantiates the cohort in the cohort table. Optionally, the inclusion rule statistics are computed and stored in the inclusion rule statistics tables described in createCohortTable).

Usage

instantiateCohort(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  oracleTempSchema = NULL,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort",
  baseUrl = NULL,
  cohortJson = NULL,
  cohortSql = NULL,
  cohortId = NULL,
  generateInclusionStats = FALSE,
  resultsDatabaseSchema = cohortDatabaseSchema,
  cohortInclusionTable = paste0(cohortTable, "_inclusion"),
  cohortInclusionResultTable = paste0(cohortTable, "_inclusion_result"),
  cohortInclusionStatsTable = paste0(cohortTable, "_inclusion_stats"),
  cohortSummaryStatsTable = paste0(cohortTable, "_summary_stats")
)

Arguments

connectionDetails
An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package. Can be left NULL if connection is provided.

connection
An object of type connection as created using the connect function in the DatabaseConnector package. Can be left NULL if connectionDetails is provided, in which case a new connection will be opened at the start of the function, and closed when the function finishes.

cdmDatabaseSchema
Schema name where your patient-level data in OMOP CDM format resides. Note that for SQL Server, this should include both the database and schema name, for example 'cdm_data.dbo'.

oracleTempSchema
Should be used in Oracle to specify a schema where the user has write privileges for storing temporary tables.

cohortDatabaseSchema
Schema name where your cohort table resides. Note that for SQL Server, this should include both the database and schema name, for example 'scratch.dbo'.

cohortTable
Name of the cohort table.

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI". Needn't be provided if cohortJson and cohortSql are provided.

cohortJson
A character string containing the JSON of a cohort definition. Needn't be provided if baseUrl and cohortId are provided.

cohortSql
The OHDSI SQL representation of the same cohort definition. Needn't be provided if baseUrl and cohortId are provided.

cohortId
The cohort ID used to reference the cohort in the cohort table.

generateInclusionStats
Compute and store inclusion rule statistics?

resultsDatabaseSchema
Schema name where the statistics tables reside. Note that for SQL Server, this should include both the database and schema name, for example 'scratch.dbo'.

cohortInclusionTable
Name of the inclusion table, one of the tables for storing inclusion rule statistics.

cohortInclusionResultTable
Name of the inclusion result table, one of the tables for storing inclusion rule statistics.

cohortInclusionStatsTable
Name of the inclusion stats table, one of the tables for storing inclusion rule statistics.

cohortSummaryStatsTable
Name of the summary stats table, one of the tables for storing inclusion rule statistics.

[Package CohortDiagnostics version 2.0.0 Index]