Launch the Diagnostics Explorer Shiny app

Launch the Diagnostics Explorer Shiny app

launchDiagnosticsExplorer(
  sqliteDbPath = "MergedCohortDiagnosticsData.sqlite",
  connectionDetails = NULL,
  shinyConfigPath = NULL,
  resultsDatabaseSchema = NULL,
  vocabularyDatabaseSchema = NULL,
  vocabularyDatabaseSchemas = resultsDatabaseSchema,
  tablePrefix = "",
  cohortTableName = "cohort",
  databaseTableName = "database",
  aboutText = NULL,
  runOverNetwork = FALSE,
  port = 80,
  makePublishable = FALSE,
  publishDir = file.path(getwd(), "DiagnosticsExplorer"),
  overwritePublishDir = FALSE,
  launch.browser = FALSE,
  enableAnnotation = TRUE
)

Arguments

sqliteDbPath

Path to merged sqlite file. See createMergedResultsFile to create file.

connectionDetails

An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package, specifying how to connect to the server where the CohortDiagnostics results have been uploaded using the uploadResults function.

shinyConfigPath

Path to shiny yml configuration file (use instead of sqliteDbPath or connectionDetails object)

resultsDatabaseSchema

The schema on the database server where the CohortDiagnostics results have been uploaded.

vocabularyDatabaseSchema

(Deprecated) Please use vocabularyDatabaseSchemas.

vocabularyDatabaseSchemas

(optional) A list of one or more schemas on the database server where the vocabulary tables are located. The default value is the value of the resultsDatabaseSchema. We can provide a list of vocabulary schema that might represent different versions of the OMOP vocabulary tables. It allows us to compare the impact of vocabulary changes on Diagnostics. Not supported with an sqlite database.

tablePrefix

(Optional) string to insert before table names (e.g. "cd_") for database table names

cohortTableName

(Optional) if cohort table name differs from the standard - cohort (ignores prefix if set)

databaseTableName

(Optional) if database table name differs from the standard - database (ignores prefix if set)

aboutText

Text (using HTML markup) that will be displayed in an About tab in the Shiny app. If not provided, no About tab will be shown.

runOverNetwork

(optional) Do you want the app to run over your network?

port

(optional) Only used if runOverNetwork = TRUE.

makePublishable

(Optional) copy data files to make app publishable to posit connect/shinyapp.io

publishDir

If make publishable is true - the directory that the shiny app is copied to

overwritePublishDir

(Optional) If make publishable is true - overwrite the directory for publishing

launch.browser

Should the app be launched in your default browser, or in a Shiny window. Note: copying to clipboard will not work in a Shiny window.

enableAnnotation

Enable annotation functionality in shiny app

Details

Launches a Shiny app that allows the user to explore the diagnostics