launchDiagnosticsExplorer {CohortDiagnostics}	R Documentation

Launch the Diagnostics Explorer Shiny app

Description

Launch the Diagnostics Explorer Shiny app

Usage

launchDiagnosticsExplorer(
  dataFolder = "data",
  connectionDetails = NULL,
  resultsDatabaseSchema = NULL,
  vocabularyDatabaseSchema = resultsDatabaseSchema,
  cohortBaseUrl = "https://atlas.ohdsi.org/#/cohortdefinition/",
  conceptBaseUrl = "https://athena.ohdsi.org/search-terms/terms/",
  runOverNetwork = FALSE,
  port = 80,
  launch.browser = FALSE
)

Arguments

dataFolder
A folder where the premerged file is stored. Use the preMergeDiagnosticsFiles function to generate this file.

connectionDetails
An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package, specifying how to connect to the server where the CohortDiagnostics results have been uploaded using the uploadResults function.

resultsDatabaseSchema
The schema on the database server where the CohortDiagnostics results have been uploaded.

vocabularyDatabaseSchema
The schema on the database server where the vocabulary tables are located.

cohortBaseUrl
The base URL for constructing linkouts to an ATLAS instance, using the webApiCohortId in the cohortsToCreate file. If NULL, no linkouts will be created.

conceptBaseUrl
The base URL for constructing linkouts to an Athena instance, using the concept ID.

runOverNetwork
(optional) Do you want the app to run over your network?

port
(optional) Only used if runOverNetwork = TRUE.

launch.browser
Should the app be launched in your default browser, or in a Shiny window. Note: copying to clipboard will not work in a Shiny window.

Details

Launches a Shiny app that allows the user to explore the diagnostics

[Package CohortDiagnostics version 2.0.0 Index]